// xa-usecases/projects/@create-ip-network/ut-create-network-config/src/app/models/network-options.model.ts
export class NetworkOptions {


        NetworkType: string;
        ProviderIndependent: boolean;
        IPType: string;
        IPQuantity: string;
        IPNetwork: string;
        IPAuthority: string;
        IPComment: string;
        AffectedDevices:string;
        AffectedCategory:string;
        VLANID: string;
       
}